Background/Sprites/Scripts
Updated background to be the part that moves to fix the jitter of the spacecraft. Background is now one large gradient from light blue to black. Updated spacecraft sprites to be a *little* more accurate and detailed. Updated scripts: player no longer moves with update. BackgroundMove script allows for background to move slowly down and increase speed over time. Also there is a pause boolean that will allow for pausing of the movement based on... different.... idk yet... benchmarks? Like each progress between stages will have a deadline to meet for certain questions. Then each stage will be like a "BOSS" trivia and pause the movement while in that "BOSS FIGHT" etc. Added mock-up drawings of ideas to a new folder (Blueprints) to document my progress.